Add TenantPipelineInitializer
